dodger: factored out keyboard and fallingball
